[tests,ci] Put the acceptance tests behind an e2e build tag
Every test in internal/provider was in one build. 65 of the 131 need a live
NetBird deployment, and the harness that starts one shells out to `docker compose
up -d` from inside the test binary — so `go test ./...` reached for Docker, and CI
stood that deployment up six times, once per Terraform CLI version, to run the 66
tests that need nothing at all.

Split them on the `e2e` build tag. The harness moves to e2e_harness_test.go and
each resource's acceptance tests to <resource>_acc_test.go, all carrying the
constraint; the conversion and matching tests stay where they are. Without the tag
none of it is compiled, which is a stronger guarantee than a runtime switch: there
is no Docker client and no compose invocation in the default test binary to
misfire, and `go test ./...` answers in under a second.

This is pure motion. The 181 declarations are byte-identical and only relocated,
and the two builds partition the tests exactly:

	go test -list          ./... ->  66 tests   (was 131)
	go test -tags e2e -list ./... -> 131 tests   (identical to before)

A helper follows the tests that use it, and nothing had to be duplicated to make
that work: of the shared helpers only valPtr is used from both halves, so it stays
untagged. Config builders turned out to be called exclusively from acceptance
tests, so each moves with them.

Split the workflows to match. test.yml builds, lints, checks generated docs and
runs the fast suite — no Docker, no Terraform CLI, about a minute. e2e.yml runs
the acceptance suite per Terraform CLI version. Lint runs twice, with and without
the tag, so the tagged half is covered too.

Move the CLI matrix to the current line and the three before it. It was pinned to
1.7 through 1.12 while 1.15 is current, so the version most users run was the one
never exercised. A pull request runs the current line and main runs all four,
since the deployment is the cost and a break against a four-line-old CLI matters
less than one against the current release.

`make testacc` grows the tag with it. Without that it would keep passing while
testing nothing, which is worse than failing.
