+++ .github/workflows/e2e.yml
@@ -1,0 +1,68 @@
+# End-to-end tests: the acceptance suite, against a NetBird deployment the tests
+# stand up themselves. Everything that runs without a deployment lives in
+# test.yml, which answers in about a minute.
+#
+# The Terraform CLI matrix is the one thing worth standing a deployment up more
+# than once for, since it is the provider's protocol surface that differs between
+# CLI versions. Paying for a deployment per version on every pull request buys
+# little, so a pull request runs the current CLI and the full matrix runs on main,
+# where a version-specific break still surfaces before a release.
+name: End-to-end
+
+on:
+  pull_request:
+    paths-ignore:
+      - 'README.md'
+  push:
+    branches:
+      - main
+    paths-ignore:
+      - 'README.md'
+  # For running the full matrix on a branch before merging it.
+  workflow_dispatch:
+    inputs:
+      full_matrix:
+        description: 'Run every supported Terraform version'
+        type: boolean
+        default: false
+
+permissions:
+  contents: read
+
+jobs:
+  acceptance:
+    name: Acceptance tests
+    runs-on: ubuntu-latest
+    timeout-minutes: 30
+    strategy:
+      fail-fast: false
+      matrix:
+        # The current CLI line and the three before it on main and on a dispatch
+        # that asks for them; the current line alone everywhere else. Each entry is
+        # a wildcard, so a patch release is picked up without a change here — only
+        # a new minor line is worth editing for.
+        terraform: ${{ fromJSON((github.event_name == 'push' || (github.event_name == 'workflow_dispatch' && inputs.full_matrix)) && '["1.12.*","1.13.*","1.14.*","1.15.*"]' || '["1.15.*"]') }}
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+      - uses: actions/setup-go@4b73464bb391d4059bd26b0524d20df3927bd417 # v6.3.0
+        with:
+          go-version-file: 'go.mod'
+          cache: true
+      - uses: hashicorp/setup-terraform@b9cd54a3c349d3f38e8881555d616ced269862dd # v3.1.2
+        with:
+          terraform_version: ${{ matrix.terraform }}
+          terraform_wrapper: false
+      - run: go mod download
+      # -tags e2e is what compiles the harness and the acceptance tests in; TF_ACC
+      # is terraform-plugin-testing's own switch. Both are needed, and neither on
+      # its own runs anything against a deployment.
+      - env:
+          TF_ACC: "1"
+        run: go test -v -tags e2e -timeout 20m -cover -coverprofile=coverage.out ./internal/provider/
+        timeout-minutes: 20
+      - name: Coverage summary
+        if: always()
+        run: |
+          if [ -f coverage.out ]; then
+            go tool cover -func=coverage.out | tail -n 1
+          fi
