+++ internal/provider/agent_network_acc_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package provider
 
 import (
