+++ internal/provider/dns_zone_acc_test.go
@@ -1,0 +1,112 @@
+//go:build e2e
+
+package provider
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func Test_DNSZone_Create(t *testing.T) {
+	rName := "z" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	rNameFull := "netbird_dns_zone." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				ResourceName: rName,
+				Config:       testDNSZoneResource(rName, "test.local", true, false, `["group-all"]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					resource.TestCheckResourceAttr(rNameFull, "domain", "test.local"),
+					resource.TestCheckResourceAttr(rNameFull, "enabled", "true"),
+					resource.TestCheckResourceAttr(rNameFull, "enable_search_domain", "false"),
+				),
+			},
+			{
+				ResourceName:      rNameFull,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func Test_DNSZone_Update(t *testing.T) {
+	rName := "z" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	rNameFull := "netbird_dns_zone." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				ResourceName: rName,
+				Config:       testDNSZoneResource(rName, "test.local", true, false, `["group-all"]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					resource.TestCheckResourceAttr(rNameFull, "domain", "test.local"),
+					resource.TestCheckResourceAttr(rNameFull, "enabled", "true"),
+					resource.TestCheckResourceAttr(rNameFull, "enable_search_domain", "false"),
+				),
+			},
+			{
+				ResourceName: rName,
+				Config:       testDNSZoneResource(rName, "test.local", false, true, `["group-all"]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					resource.TestCheckResourceAttr(rNameFull, "domain", "test.local"),
+					resource.TestCheckResourceAttr(rNameFull, "enabled", "false"),
+					resource.TestCheckResourceAttr(rNameFull, "enable_search_domain", "true"),
+				),
+			},
+		},
+	})
+}
+
+func Test_DNSZone_DataSource(t *testing.T) {
+	rName := "z" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	dsNameFull := "data.netbird_dns_zone." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: testDNSZoneResourceAndDataSource(rName, "datasource.local", true, false, `["group-all"]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(dsNameFull, "id"),
+					resource.TestCheckResourceAttr(dsNameFull, "name", rName),
+					resource.TestCheckResourceAttr(dsNameFull, "domain", "datasource.local"),
+					resource.TestCheckResourceAttr(dsNameFull, "enabled", "true"),
+					resource.TestCheckResourceAttr(dsNameFull, "enable_search_domain", "false"),
+				),
+			},
+		},
+	})
+}
+
+func testDNSZoneResource(name, domain string, enabled, enableSearchDomain bool, distributionGroups string) string {
+	return fmt.Sprintf(`
+resource "netbird_dns_zone" "%s" {
+  name                 = "%s"
+  domain               = "%s"
+  enabled              = %t
+  enable_search_domain = %t
+  distribution_groups  = %s
+}
+`, name, name, domain, enabled, enableSearchDomain, distributionGroups)
+}
+
+func testDNSZoneResourceAndDataSource(name, domain string, enabled, enableSearchDomain bool, distributionGroups string) string {
+	return testDNSZoneResource(name, domain, enabled, enableSearchDomain, distributionGroups) + fmt.Sprintf(`
+data "netbird_dns_zone" "%s" {
+  name = netbird_dns_zone.%s.name
+}
+`, name, name)
+}
