+++ internal/provider/group_acc_test.go
@@ -1,0 +1,105 @@
+//go:build e2e
+
+package provider
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+)
+
+func Test_Group_Create(t *testing.T) {
+	rName := "g" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	rNameFull := "netbird_group." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				ResourceName: rName,
+				Config:       testGroupResource(rName, `[]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					func(s *terraform.State) error {
+						gID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
+						group, err := testClient().Groups.Get(context.Background(), gID)
+						if err != nil {
+							return err
+						}
+						if group.Name != rName {
+							return fmt.Errorf("Group name mismatch, expected %s, found %s on management server", rName, group.Name)
+						}
+						return nil
+					},
+				),
+			},
+		},
+	})
+}
+
+func Test_Group_Update(t *testing.T) {
+	rName := "g" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	rNameFull := "netbird_group." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		CheckDestroy: func(s *terraform.State) error {
+			groups, err := testClient().Groups.List(context.Background())
+			if err != nil {
+				return err
+			}
+			for _, g := range groups {
+				if g.Name == rName {
+					return fmt.Errorf("Group not deleted")
+				}
+			}
+			return nil
+		},
+		Steps: []resource.TestStep{
+			{
+				ResourceName: rName,
+				Config:       testGroupResource(rName, `[]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+				),
+			},
+			{
+				ResourceName: rName,
+				Config:       testGroupResource(rName, `["peer1"]`),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					resource.TestCheckResourceAttr(rNameFull, "peers.#", "1"),
+					resource.TestCheckResourceAttr(rNameFull, "peers.0", "peer1"),
+					func(s *terraform.State) error {
+						gID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
+						group, err := testClient().Groups.Get(context.Background(), gID)
+						if err != nil {
+							return err
+						}
+						if len(group.Peers) != 1 {
+							return fmt.Errorf("Group Peers not updated in management")
+						}
+						if group.Peers[0].Id != "peer1" {
+							return fmt.Errorf("Group Peers incorrect")
+						}
+						return nil
+					},
+				),
+			},
+		},
+	})
+}
+
+func testGroupResource(rName, peers string) string {
+	return fmt.Sprintf(`resource "netbird_group" "%s" {
+	name = "%s"
+	peers = %s
+}`, rName, rName, peers)
+}
