+++ internal/provider/group_test.go
@@ -2,15 +2,11 @@ package provider
 
 import (
 	"context"
-	"fmt"
 	"reflect"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/netbirdio/netbird/shared/management/http/api"
 )
 
@@ -85,96 +81,4 @@ func Test_groupAPIToTerraform(t *testing.T) {
 			t.Fatalf("Expected:\n%#v\nFound:\n%#v", c.expected, out)
 		}
 	}
-}
-
-func Test_Group_Create(t *testing.T) {
-	rName := "g" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
-	rNameFull := "netbird_group." + rName
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testEnsureManagementRunning(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				ResourceName: rName,
-				Config:       testGroupResource(rName, `[]`),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-					func(s *terraform.State) error {
-						gID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
-						group, err := testClient().Groups.Get(context.Background(), gID)
-						if err != nil {
-							return err
-						}
-						if group.Name != rName {
-							return fmt.Errorf("Group name mismatch, expected %s, found %s on management server", rName, group.Name)
-						}
-						return nil
-					},
-				),
-			},
-		},
-	})
-}
-
-func Test_Group_Update(t *testing.T) {
-	rName := "g" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
-	rNameFull := "netbird_group." + rName
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testEnsureManagementRunning(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		CheckDestroy: func(s *terraform.State) error {
-			groups, err := testClient().Groups.List(context.Background())
-			if err != nil {
-				return err
-			}
-			for _, g := range groups {
-				if g.Name == rName {
-					return fmt.Errorf("Group not deleted")
-				}
-			}
-			return nil
-		},
-		Steps: []resource.TestStep{
-			{
-				ResourceName: rName,
-				Config:       testGroupResource(rName, `[]`),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-				),
-			},
-			{
-				ResourceName: rName,
-				Config:       testGroupResource(rName, `["peer1"]`),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-					resource.TestCheckResourceAttr(rNameFull, "peers.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "peers.0", "peer1"),
-					func(s *terraform.State) error {
-						gID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
-						group, err := testClient().Groups.Get(context.Background(), gID)
-						if err != nil {
-							return err
-						}
-						if len(group.Peers) != 1 {
-							return fmt.Errorf("Group Peers not updated in management")
-						}
-						if group.Peers[0].Id != "peer1" {
-							return fmt.Errorf("Group Peers incorrect")
-						}
-						return nil
-					},
-				),
-			},
-		},
-	})
-}
-
-func testGroupResource(rName, peers string) string {
-	return fmt.Sprintf(`resource "netbird_group" "%s" {
-	name = "%s"
-	peers = %s
-}`, rName, rName, peers)
 }
