+++ internal/provider/peer_acc_test.go
@@ -1,0 +1,89 @@
+//go:build e2e
+
+package provider
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
+)
+
+func Test_Peer_Create(t *testing.T) {
+	rName := "p" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	rNameFull := "netbird_peer." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				ResourceName: rName,
+				Config:       testPeerResource(rName, `peer2`, rName),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					func(s *terraform.State) error {
+						pID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
+						peer, err := testClient().Peers.Get(context.Background(), pID)
+						if err != nil {
+							return err
+						}
+
+						if peer.Name != rName {
+							return fmt.Errorf("Peer name mismatch, expected %s, found %s on management server", rName, peer.Name)
+						}
+
+						return nil
+					},
+				),
+			},
+		},
+	})
+}
+
+func Test_Peer_Update(t *testing.T) {
+	rName := "p" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
+	rNameFull := "netbird_peer." + rName
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testEnsureManagementRunning(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				ResourceName: rName,
+				Config:       testPeerResource(rName, `peer3`, "meow"),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+				),
+			},
+			{
+				ResourceName: rName,
+				Config:       testPeerResource(rName, `peer3`, rName),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(rNameFull, "id"),
+					resource.TestCheckResourceAttr(rNameFull, "name", rName),
+					func(s *terraform.State) error {
+						pID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
+						peer, err := testClient().Peers.Get(context.Background(), pID)
+						if err != nil {
+							return err
+						}
+						if peer.Name != rName {
+							return fmt.Errorf("Peer name mismatch, expected %s, found %s on management server", rName, peer.Name)
+						}
+						return nil
+					},
+				),
+			},
+		},
+	})
+}
+
+func testPeerResource(rName, id, name string) string {
+	return fmt.Sprintf(`resource "netbird_peer" "%s" {
+	id = "%s"
+	name = "%s"
+}`, rName, id, name)
+}
