+++ internal/provider/policy_test.go
@@ -2,7 +2,6 @@ package provider
 
 import (
 	"context"
-	"fmt"
 	"reflect"
 	"regexp"
 	"strconv"
@@ -10,9 +9,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/netbirdio/netbird/shared/management/http/api"
 )
 
@@ -377,267 +373,4 @@ func Test_portRegex(t *testing.T) {
 			t.Fatalf("Exepected %s not to match regex", c)
 		}
 	}
-}
-
-func Test_Policy_Create_Groups(t *testing.T) {
-	rName := "po" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
-	rNameFull := "netbird_policy." + rName
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testEnsureManagementRunning(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				ResourceName: rName,
-				Config:       testPolicyResourceGroups(rName, rName, "desc", "accept", "udp", "group-all", "group-notall", "443"),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-					resource.TestCheckResourceAttr(rNameFull, "description", "desc"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.action", "accept"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.ports.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.ports.0", "443"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.sources.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.sources.0", "group-all"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destinations.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destinations.0", "group-notall"),
-					func(s *terraform.State) error {
-						pID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
-						policy, err := testClient().Policies.Get(context.Background(), pID)
-						if err != nil {
-							return err
-						}
-
-						return matchPairs(map[string][]any{
-							"Name":                    {rName, policy.Name},
-							"Description":             {"desc", policy.Description},
-							"Rules.#":                 {int(1), len(policy.Rules)},
-							"Rules[0].Action":         {"accept", string(policy.Rules[0].Action)},
-							"Rules[0].Ports.#":        {int(1), len(*policy.Rules[0].Ports)},
-							"Rules[0].Ports.0":        {"443", (*policy.Rules[0].Ports)[0]},
-							"Rules[0].Sources.#":      {int(1), len(*policy.Rules[0].Sources)},
-							"Rules[0].Sources.0":      {"group-all", (*policy.Rules[0].Sources)[0].Id},
-							"Rules[0].Destinations.#": {int(1), len(*policy.Rules[0].Destinations)},
-							"Rules[0].Destinations.0": {"group-notall", (*policy.Rules[0].Destinations)[0].Id},
-						})
-					},
-				),
-			},
-		},
-	})
-}
-
-func Test_Policy_Create_Resources(t *testing.T) {
-	rName := "po" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
-	rNameFull := "netbird_policy." + rName
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testEnsureManagementRunning(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				ResourceName: rName,
-				Config:       testPolicyResourceResources(rName, rName, "desc", "accept", "udp", "resource2", "subnet", "resource1", "domain", "1000", "2000"),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-					resource.TestCheckResourceAttr(rNameFull, "description", "desc"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.action", "accept"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.ports.#", "0"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.sources.#", "0"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.source_resource.id", "resource2"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.source_resource.type", "subnet"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destination_resource.id", "resource1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destination_resource.type", "domain"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destinations.#", "0"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.port_ranges.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.port_ranges.0.start", "1000"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.port_ranges.0.end", "2000"),
-					func(s *terraform.State) error {
-						pID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
-						policy, err := testClient().Policies.Get(context.Background(), pID)
-						if err != nil {
-							return err
-						}
-
-						return matchPairs(map[string][]any{
-							"Name":                            {rName, policy.Name},
-							"Description":                     {"desc", policy.Description},
-							"Rules.#":                         {int(1), len(policy.Rules)},
-							"Rules[0].Action":                 {"accept", string(policy.Rules[0].Action)},
-							"Rules[0].Sources.#":              {nil, policy.Rules[0].Sources},
-							"Rules[0].Destinations.#":         {nil, policy.Rules[0].Destinations},
-							"Rules[0].SourceResource.ID":      {"resource2", policy.Rules[0].SourceResource.Id},
-							"Rules[0].DestinationResource.ID": {"resource1", policy.Rules[0].DestinationResource.Id},
-							"Rules[0].PortRanges.#":           {int(1), len(*policy.Rules[0].PortRanges)},
-							"Rules[0].PortRanges.0.Start":     {int(1000), (*policy.Rules[0].PortRanges)[0].Start},
-							"Rules[0].PortRanges.0.End":       {int(2000), (*policy.Rules[0].PortRanges)[0].End},
-						})
-					},
-				),
-			},
-		},
-	})
-
-}
-
-func Test_Policy_Update_Groups(t *testing.T) {
-	rName := "po" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
-	rNameFull := "netbird_policy." + rName
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testEnsureManagementRunning(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				ResourceName: rName,
-				Config:       testPolicyResourceGroups(rName, rName, "desc", "accept", "udp", "group-all", "group-notall", "443"),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-				),
-			},
-			{
-				ResourceName: rName,
-				Config:       testPolicyResourceGroups(rName, rName, "desc-updated", "drop", "tcp", "group-notall", "group-all", "80"),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-					resource.TestCheckResourceAttr(rNameFull, "description", "desc-updated"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.action", "drop"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.ports.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.ports.0", "80"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.sources.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.sources.0", "group-notall"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destinations.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destinations.0", "group-all"),
-					func(s *terraform.State) error {
-						pID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
-						policy, err := testClient().Policies.Get(context.Background(), pID)
-						if err != nil {
-							return err
-						}
-						return matchPairs(map[string][]any{
-							"Name":                    {rName, policy.Name},
-							"Description":             {"desc-updated", policy.Description},
-							"Rules.#":                 {int(1), len(policy.Rules)},
-							"Rules[0].Action":         {"drop", string(policy.Rules[0].Action)},
-							"Rules[0].Sources.#":      {int(1), len(*policy.Rules[0].Sources)},
-							"Rules[0].Sources.0":      {"group-notall", (*policy.Rules[0].Sources)[0].Id},
-							"Rules[0].Destinations.#": {int(1), len(*policy.Rules[0].Destinations)},
-							"Rules[0].Destinations.0": {"group-all", (*policy.Rules[0].Destinations)[0].Id},
-							"Rules[0].Ports.#":        {int(1), len(*policy.Rules[0].Ports)},
-							"Rules[0].Ports.0":        {"80", (*policy.Rules[0].Ports)[0]},
-						})
-					},
-				),
-			},
-		},
-	})
-}
-
-func Test_Policy_Update_Resources(t *testing.T) {
-	rName := "po" + acctest.RandStringFromCharSet(10, acctest.CharSetAlpha)
-	rNameFull := "netbird_policy." + rName
-	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testEnsureManagementRunning(t) },
-		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				ResourceName: rName,
-				Config:       testPolicyResourceGroups(rName, rName, "desc", "accept", "udp", "group-all", "group-notall", "80"),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-				),
-			},
-			{
-				ResourceName: rName,
-				Config:       testPolicyResourceResources(rName, rName, "desc", "accept", "udp", "resource2", "subnet", "resource1", "domain", "1", "100"),
-				Check: resource.ComposeAggregateTestCheckFunc(
-					resource.TestCheckResourceAttrSet(rNameFull, "id"),
-					resource.TestCheckResourceAttr(rNameFull, "name", rName),
-					resource.TestCheckResourceAttr(rNameFull, "rule.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.action", "accept"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.ports.#", "0"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.sources.#", "0"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.source_resource.id", "resource2"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.source_resource.type", "subnet"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destination_resource.id", "resource1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destination_resource.type", "domain"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.destinations.#", "0"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.port_ranges.#", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.port_ranges.0.start", "1"),
-					resource.TestCheckResourceAttr(rNameFull, "rule.0.port_ranges.0.end", "100"),
-					func(s *terraform.State) error {
-						pID := s.RootModule().Resources[rNameFull].Primary.Attributes["id"]
-						policy, err := testClient().Policies.Get(context.Background(), pID)
-						if err != nil {
-							return err
-						}
-						return matchPairs(map[string][]any{
-							"Name":                            {rName, policy.Name},
-							"Description":                     {"desc", policy.Description},
-							"Rules.#":                         {int(1), len(policy.Rules)},
-							"Rules[0].Action":                 {"accept", string(policy.Rules[0].Action)},
-							"Rules[0].Sources.#":              {nil, policy.Rules[0].Sources},
-							"Rules[0].Destinations.#":         {nil, policy.Rules[0].Destinations},
-							"Rules[0].SourceResource.ID":      {"resource2", policy.Rules[0].SourceResource.Id},
-							"Rules[0].DestinationResource.ID": {"resource1", policy.Rules[0].DestinationResource.Id},
-							"Rules[0].PortRanges.#":           {int(1), len(*policy.Rules[0].PortRanges)},
-							"Rules[0].PortRanges.0.Start":     {int(1), (*policy.Rules[0].PortRanges)[0].Start},
-							"Rules[0].PortRanges.0.End":       {int(100), (*policy.Rules[0].PortRanges)[0].End},
-						})
-					},
-				),
-			},
-		},
-	})
-}
-
-func testPolicyResourceGroups(rName, name, description, rAction, rProt, rSource, rDest, port string) string {
-	return fmt.Sprintf(`resource "netbird_policy" "%s" {
-	name    = "%s"
-	description = "%s"
-	enabled = true
-
-	rule {
-		action        = "%s"
-		bidirectional = true
-		enabled       = true
-		protocol      = "%s"
-		name          = "%s"
-		sources       = ["%s"]
-		destinations  = ["%s"]
-		ports         = ["%s"]
-	}
-}`, rName, name, description, rAction, rProt, name, rSource, rDest, port)
-}
-
-func testPolicyResourceResources(rName, name, description, rAction, rProt, rSourceID, rSourceType, rDestID, rDestType, pStart, pEnd string) string {
-	return fmt.Sprintf(`resource "netbird_policy" "%s" {
-	name        = "%s"
-	description = "%s"
-	enabled     = true
-
-	rule {
-		action                = "%s"
-		bidirectional         = true
-		enabled               = true
-		protocol              = "%s"
-		name                  = "%s"
-		source_resource       = {
-			id = "%s"
-			type = "%s"
-		}
-		destination_resource  = {
-			id = "%s"
-			type = "%s"
-		}
-		port_ranges = [
-			{
-				start = %s
-				end   = %s
-			}
-		]
-	}
-}`, rName, name, description, rAction, rProt, name, rSourceID, rSourceType, rDestID, rDestType, pStart, pEnd)
 }
