+++ internal/provider/reverse_proxy_acc_test.go
@@ -1,3 +1,5 @@
+//go:build e2e
+
 package provider
 
 import (
